+++ packages/instantsearch-ui-components/src/components/index.ts
@@ -16,6 +16,7 @@ export * from './chat/ChatPromptSuggestions';
 export * from './chat/ChatToggleButton';
 export * from './chat/tools/SearchIndexTool';
 export * from './chat/icons';
+export * from './chat/tools/DisplayResultsTool';
 export * from './chat/types';
 export * from './FrequentlyBoughtTogether';
 export * from './Highlight';