+++ packages/react-instantsearch/src/widgets/Autocomplete.tsx
@@ -15,6 +15,7 @@ import {
   getPromptSuggestionHits,
   isPromptSuggestion,
 } from 'instantsearch-ui-components';
+import { isChatBusy, openChat } from 'instantsearch.js/es/lib/chat';
 import { warn } from 'instantsearch.js/es/lib/utils';
 import React, {
   createElement,
@@ -742,14 +743,10 @@ function InnerAutocomplete<TItem extends BaseHit = BaseHit>({
       userOnSelect ??
       (({ item, query, setQuery, url }) => {
         if (isPromptSuggestion(item)) {
-          const chatRenderStateWithFocus = chatRenderState as
-            | (Partial<ChatRenderState> & { focusInput?: () => void })
-            | undefined;
-
-          if (chatRenderStateWithFocus) {
-            chatRenderStateWithFocus.setOpen?.(true);
-            chatRenderStateWithFocus.focusInput?.();
-            chatRenderStateWithFocus.sendMessage?.({ text: item.prompt });
+          if (chatRenderState) {
+            if (openChat(chatRenderState, { message: item.prompt })) {
+              setQuery('');
+            }
             return;
           }
 
@@ -841,7 +838,7 @@ function InnerAutocomplete<TItem extends BaseHit = BaseHit>({
     );
   }
 
-  indicesForPanel.forEach(({ indexId, indexName, hits }) => {
+  indicesForPanel.forEach(({ indexId, indexName, hits, sendEvent }) => {
     let elementId = indexName;
     if (indexName === showQuerySuggestions?.indexName) {
       elementId = 'suggestions';
@@ -869,10 +866,16 @@ function InnerAutocomplete<TItem extends BaseHit = BaseHit>({
           __indexName: indexId,
         }))}
         getItemProps={getItemProps}
+        sendEvent={sendEvent}
         classNames={currentIndexConfig.classNames}
       />
     );
   });
+
+  const handleCancel = () => {
+    setIsModalOpen(false);
+    setIsOpen(false);
+  };
 
   const searchBoxContent = (
     <AutocompleteSearch
@@ -886,23 +889,39 @@ function InnerAutocomplete<TItem extends BaseHit = BaseHit>({
       }}
       query={resolvedQuery}
       isSearchStalled={isSearchStalled}
+      onCancel={() => {
+        if (isDetached) {
+          handleCancel();
+        }
+      }}
+      isDetached={isDetached}
+      submitTitle={
+        isDetached ? translations.detachedCancelButtonText : undefined
+      }
       onAiModeClick={
         aiMode
           ? () => {
-              if (chatRenderState) {
-                chatRenderState.setOpen?.(true);
-                if (resolvedQuery.trim()) {
-                  chatRenderState.sendMessage?.({ text: resolvedQuery });
-                }
+              setIsOpen(false);
+              if (isDetached) {
+                setIsModalOpen(false);
+              }
+              if (openChat(chatRenderState, { message: resolvedQuery })) {
+                refineSearchBox('');
+                refineAutocomplete('');
               }
             }
           : undefined
       }
+      aiModeButtonDisabled={aiMode ? isChatBusy(chatRenderState) : undefined}
+      classNames={classNames}
     />
   );
 
   const panelContent = (
-    <AutocompletePanel {...getPanelProps()}>
+    <AutocompletePanel
+      {...getPanelProps()}
+      classNames={{ root: classNames?.panel, open: classNames?.panelOpen, layout: classNames?.panelLayout }}
+    >
       {PanelComponent ? (
         <PanelComponent
           elements={elements}
@@ -951,21 +970,13 @@ function InnerAutocomplete<TItem extends BaseHit = BaseHit>({
         {isModalOpen && (
           <AutocompleteDetachedOverlay
             classNames={classNames}
-            onClose={() => {
-              setIsModalOpen(false);
-              setIsOpen(false);
-            }}
+            onClose={handleCancel}
           >
             <AutocompleteDetachedContainer
               classNames={detachedContainerClassNames}
             >
               <AutocompleteDetachedFormContainer
                 classNames={classNames}
-                onCancel={() => {
-                  setIsModalOpen(false);
-                  setIsOpen(false);
-                }}
-                translations={translations}
               >
                 {searchBoxContent}
               </AutocompleteDetachedFormContainer>